requirements.txt
crewai==0.1.32
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
urllib3==1.26.18 